backend/appointments/requirements.txt
Django==4.0.4
PyYAML==5.3.1
uritemplate==3.0.1
psycopg2-binary==2.9.3
django-environ==0.8.1
requests==2.22.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
pytz==2021.1